posting image on controller
