+++ models/order.model.js
@@ -1,30 +1,2 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-let order = new Schema({
-    user:{
-        type:Schema.Types.ObjectId,
-        ref:"User"
-    },
-    products:[
-        {
-            type:Schema.Types.ObjectId,
-            ref:"Product"
-        }
-    ],
-    address:{
-        city:String, // Thanh pho
-        town:String, // Quan
-        ward:String, // Phuong
-        street:String // Dia chi cu the tu viet
-    },
-    total: Number,
-    shipper:{
-        type:Schema.Types.ObjectId,
-        ref:"Shipper"
-    },
-    status:{
-        type:String,
-        enum:["receiving, delivering, delivered, canceled"]
-    }
-},{timestamps:true});
-module.exports = mongoose.model("Order",order,"Order");
+var mongoose = require("mongoose");
+var Schema = mongoose.Schema;
