+++ routes/product.route.js
@@ -1,11 +1,13 @@
 let express = require("express");
 let router = express.Router();
 let productController = require("../controllers/product.controller");
+let upload = require("../controllers/upload.controller");
 
+let productUpload = upload.single('images');
 
 router.get("/",productController.getProducts);
 router.get("/:id", productController.getProduct);
-router.post("/",productController.createProduct);
+router.post("/", productUpload, productController.createProduct);
 router.put("/",productController.updateProduct);
 router.delete("/:id", productController.deleteProduct);
