+++ routes/user.route.js
@@ -1,7 +1,6 @@
 let express = require("express");
 let router = express.Router();
 let userController = require("../controllers/user.controller.js");
-const { route } = require("./newfeed.route.js");
 router.get("/",userController.getUsers);
 router.get("/:id",userController.getUser);
 router.put("/",userController.updateUser);
